@@ -13,14 +13,14 @@
     "postinstall": "husky install"
   },
   "dependencies": {
-    "@0xsequence/multicall": "^0.36.2",
-    "@apollo/client": "^3.4.7",
-    "@dopex-io/sdk": "^1.0.0-alpha.63",
+    "@0xsequence/multicall": "^0.36.3",
+    "@apollo/client": "^3.5.9",
+    "@dopex-io/sdk": "^1.0.0-alpha.77",
     "@material-ui/core": "^4.11.2",
     "@material-ui/icons": "^4.11.2",
     "@material-ui/lab": "^4.0.0-alpha.58",
     "@walletconnect/web3-provider": "^1.6.5",
-    "axios": "^0.25.0",
+    "axios": "^0.26.0",
     "bignumber.js": "^9.0.1",
     "classnames": "^2.2.6",
     "date-fns": "^2.16.1",
@@ -31,44 +31,43 @@
     "fortmatic": "^2.2.1",
     "graphql": "^16.0.1",
     "lodash": "^4.17.20",
-    "next": "^12.0.3",
+    "next": "^12.1.0",
     "react": "^17.0.1",
     "react-countdown": "^2.3.2",
     "react-dom": "^17.0.1",
     "react-firebase-hooks": "^4.0.1",
     "react-hot-toast": "^2.0.0",
     "react-router-dom": "^6.0.2",
-    "react-swipeable-tab": "^0.1.1",
     "react-table": "^7.7.0",
     "react-use": "^17.2.4",
-    "recharts": "^2.0.9",
-    "sass": "^1.43.2",
+    "recharts": "^2.1.9",
+    "sass": "^1.49.8",
     "use-debounce": "^7.0.1",
     "walletlink": "^2.4.4",
     "web3modal": "^1.9.4",
     "yup": "^0.32.9"
   },
   "devDependencies": {
-    "@graphql-codegen/cli": "^2.0.1",
-    "@graphql-codegen/typescript": "^2.0.0",
-    "@graphql-codegen/typescript-operations": "^2.0.1",
-    "@graphql-codegen/typescript-react-apollo": "^3.0.0",
-    "@next/eslint-plugin-next": "^12.0.3",
+    "@graphql-codegen/cli": "^2.6.2",
+    "@graphql-codegen/typescript": "^2.4.5",
+    "@graphql-codegen/typescript-operations": "^2.3.2",
+    "@graphql-codegen/typescript-react-apollo": "^3.2.8",
+    "@next/eslint-plugin-next": "^12.1.0",
     "@types/lodash": "^4.14.170",
     "@types/react": "^17.0.11",
     "@types/react-dom": "^17.0.8",
     "@types/react-router-dom": "^5.1.7",
-    "@typescript-eslint/eslint-plugin": "^5.9.0",
-    "@typescript-eslint/parser": "^5.9.0",
+    "@typescript-eslint/eslint-plugin": "^5.12.0",
+    "@typescript-eslint/parser": "^5.12.0",
     "autoprefixer": "^10.4.0",
-    "eslint": "8.8.0",
-    "eslint-config-next": "12.0.10",
+    "eslint": "8.9.0",
+    "eslint-config-next": "12.1.0",
     "husky": "^7.0.1",
-    "lint-staged": "^12.0.1",
+    "lint-staged": "^12.3.4",
     "postcss": "^8.3.11",
     "prettier": "^2.1.1",
     "shelljs": "^0.8.4",
-    "tailwindcss": "^3.0.2",
+    "tailwindcss": "^3.0.23",
     "typescript": "^4.1.2"
   },
   "lint-staged": {
