@@ -97,6 +97,7 @@ const menuLinks = [
   { name: 'Discord', to: 'https://discord.gg/dopex' },
   { name: 'Github', to: 'https://github.com/dopex-io' },
   { name: 'Price Oracles', to: '/oracles' },
+  { name: 'Diamond Pepe NFTs', to: '/nfts/diamondpepes' },
   { name: 'Dopex NFTs', to: '/nfts' },
   { name: 'Community NFTs', to: '/nfts/community' },
 ];
@@ -122,12 +123,6 @@ export default function AppBar(props: AppBarProps) {
   const { accountAddress, connect, wrongNetwork, chainId, ensName, ensAvatar } =
     useContext(WalletContext);
   const { tokenPrices, userAssetBalances } = useContext(AssetsContext);
-  const ssovTokens = {
-    42161: ['ETH', 'DPX', 'RDPX', 'GOHM', 'GMX'],
-    421611: ['USDTMock'],
-    56: ['BNB'],
-    43114: ['AVAX'],
-  };
 
   const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
   const [anchorElSmall, setAnchorElSmall] = useState<null | HTMLElement>(null);
@@ -203,10 +198,7 @@ export default function AppBar(props: AppBarProps) {
         handleClose={handleWalletDialogClose}
       />
       <nav
-        className={cx(
-          'bg-black fixed top-0 z-10 w-full text-gray-600',
-          styles.appBar
-        )}
+        className={cx('fixed top-0 z-50 w-full text-gray-600', styles.appBar)}
       >
         <Box className="flex w-full items-center container lg:px-0 p-4 justify-between mx-auto">
           <Box className="flex items-center">
@@ -239,17 +231,16 @@ export default function AppBar(props: AppBarProps) {
           </Box>
           <Box className="flex items-center">
             <Box className="space-x-2 mr-4 hidden lg:flex">
-              {tokenPrices.map(
-                (item) =>
-                  ssovTokens[chainId].includes(item['name']) && (
-                    <PriceTag
-                      key={item.name}
-                      asset={item.name}
-                      price={item.price}
-                      change={item.change24h}
-                    />
-                  )
-              )}
+              {tokenPrices
+                .filter((item) => ['DPX', 'RDPX', 'ETH'].includes(item['name']))
+                .map((item) => (
+                  <PriceTag
+                    key={item.name}
+                    asset={item.name}
+                    price={item.price}
+                    change={item.change24h}
+                  />
+                ))}
             </Box>
             {/* {baseAssetsWithPrices ? (
               <PriceTag
