@@ -16,15 +16,19 @@ import ArrowDropDownIcon from '@material-ui/icons/ArrowDropDown';
 import { BigNumber } from 'ethers';
 import { LoaderIcon } from 'react-hot-toast';
 import { ERC20 } from '@dopex-io/sdk';
-import { WalletContext } from '../../contexts/Wallet';
-import { AssetsContext, IS_NATIVE } from '../../contexts/Assets';
+
+import { WalletContext } from 'contexts/Wallet';
+import { AssetsContext, IS_NATIVE } from 'contexts/Assets';
+
 import TokenSelector from '../TokenSelector';
 import CustomButton from '../UI/CustomButton';
 import Typography from '../UI/Typography';
-import getSymbolFromAddress from '../../utils/general/getSymbolFromAddress';
-import getUserReadableAmount from '../../utils/contracts/getUserReadableAmount';
-import getDecimalsFromSymbol from '../../utils/general/getDecimalsFromSymbol';
-import formatAmount from '../../utils/general/formatAmount';
+
+import getSymbolFromAddress from 'utils/general/getSymbolFromAddress';
+import getUserReadableAmount from 'utils/contracts/getUserReadableAmount';
+import getTokenDecimals from 'utils/general/getTokenDecimals';
+import formatAmount from 'utils/general/formatAmount';
+
 import ArrowLeftIcon from '../Icons/ArrowLeftIcon';
 import SettingsIcon from '../Icons/SettingsIcon';
 import CrossIcon from '../Icons/CrossIcon';
@@ -152,7 +156,6 @@ const ZapIn = ({
                 else setToken(ssovToken);
               }}
             />
-
             <Typography variant="h5">Zap In</Typography>
             <Tooltip
               title="Go to advanced mode"
@@ -209,7 +212,6 @@ const ZapIn = ({
               </Box>
             </Popover>
           </Box>
-
           <Box className="bg-umbra rounded-2xl flex flex-col mb-4 p-3 pr-1">
             <Box className="flex flex-row justify-between">
               <Box
@@ -237,7 +239,6 @@ const ZapIn = ({
                   <ArrowDropDownIcon className={'fill-gray-100 mr-2'} />
                 </IconButton>
               </Box>
-
               <Input
                 disableUnderline
                 id="zapInAmount"
@@ -246,8 +247,8 @@ const ZapIn = ({
                 type="number"
                 className="h-12 text-2xl text-white ml-2 mr-3 font-mono"
                 value={getUserReadableAmount(
-                  userAssetBalances[tokenName],
-                  getDecimalsFromSymbol(tokenName, chainId)
+                  userAssetBalances[tokenName.toLocaleUpperCase()],
+                  getTokenDecimals(tokenName, chainId)
                 ).toFixed(6)}
                 readOnly={true}
                 classes={{ input: 'text-right' }}
@@ -272,12 +273,11 @@ const ZapIn = ({
               </Box>
             </Box>
           </Box>
-
           <Box className={isInDialog ? 'h-[30.2rem]' : 'h-[20.2rem]'}>
             {tokenName !== '' &&
               ssovTokenName != '' &&
               ssovTokenName !== tokenName && (
-                <Box className="rounded-xl col-flex mb-4 p-3 pb-0 border border-neutral-800 w-full mb-52">
+                <Box className="rounded-xl col-flex mb-4 p-3 pb-0 border border-neutral-800 w-full">
                   <Box
                     className={
                       showSwapSteps
@@ -287,7 +287,6 @@ const ZapIn = ({
                     onClick={() => setShowSwapSteps(!showSwapSteps)}
                   >
                     <ZapIcon className="mt-1 mr-2" id="3" />
-
                     <Typography variant="h6" className="text-white font-lg">
                       1 {tokenName} ={' '}
                       {quote['toToken']
@@ -324,7 +323,6 @@ const ZapIn = ({
                       />
                     )}
                   </Box>
-
                   {showSwapSteps && (
                     <Box>
                       <Box className={'flex mb-2 mt-4'}>
@@ -355,7 +353,6 @@ const ZapIn = ({
                           </Typography>
                         </Box>
                       </Box>
-
                       <Box className={'flex mb-2 mt-2'}>
                         <Typography
                           variant="h6"
@@ -491,7 +488,6 @@ const ZapIn = ({
                 </Box>
               )}
           </Box>
-
           <Box className="rounded-xl p-4 border border-neutral-800 w-full  bg-umbra">
             <Box className="rounded-md flex mb-4 p-3 border border-neutral-800 w-full bg-neutral-800 cursor-pointer">
               <Typography variant="h6" className="text-gray-400 opacity-70">
@@ -527,7 +523,6 @@ const ZapIn = ({
           </Box>
         </Box>
       )}
-
       {isTokenSelectorVisible && (
         <Box className={isInDialog ? 'h-[52.8rem]' : 'h-[38.8rem]'}>
           <TokenSelector
@@ -535,6 +530,7 @@ const ZapIn = ({
             setOpen={setIsTokenSelectorVisible}
             setToken={setToken}
             isInDialog={isInDialog}
+            tokensToExclude={[ssovTokenName]}
           />
         </Box>
       )}
