@@ -1,4 +1,4 @@
-import { useContext, useMemo, useEffect } from 'react';
+import { useContext, useEffect } from 'react';
 import { useParams } from 'react-router-dom';
 import Head from 'next/head';
 import Box from '@material-ui/core/Box';
@@ -10,59 +10,24 @@ import ExerciseList from '../components/ExerciseList';
 import Stats from '../components/Stats';
 import PageLoader from 'components/PageLoader';
 
-import {
-  SsovContext,
-  SsovProperties,
-  SsovData,
-  UserSsovData,
-} from 'contexts/Ssov';
+import { SsovContext, SsovProvider } from 'contexts/Ssov';
+import { BnbConversionProvider } from 'contexts/BnbConversion';
 
 const Manage = () => {
-  const { asset } = useParams();
+  const { asset, type } = useParams();
   const {
-    ssovPropertiesArray,
-    ssovDataArray,
-    userSsovDataArray,
+    ssovData,
+    ssovEpochData,
+    ssovUserData,
     setSelectedSsov,
     selectedSsov,
   } = useContext(SsovContext);
 
-  const {
-    ssovProperties,
-    ssovData,
-    userSsovData,
-    index,
-  }: {
-    ssovProperties: SsovProperties;
-    ssovData: SsovData;
-    userSsovData: UserSsovData;
-    index: number;
-  } = useMemo(() => {
-    if (ssovPropertiesArray.length === 0)
-      return {
-        ssovProperties: undefined,
-        ssovData: undefined,
-        userSsovData: undefined,
-        index: 0,
-      };
-    let i = ssovPropertiesArray.findIndex((item) => item.tokenName === asset);
-    return {
-      ssovProperties: ssovPropertiesArray[i],
-      ssovData: ssovDataArray[i],
-      userSsovData: userSsovDataArray[i],
-      index: i,
-    };
-  }, [ssovPropertiesArray, asset, ssovDataArray, userSsovDataArray]);
-
   useEffect(() => {
-    setSelectedSsov(index);
-  }, [index, setSelectedSsov]);
+    setSelectedSsov({ token: asset, type: type.toUpperCase() });
+  }, [setSelectedSsov, asset, type]);
 
-  if (
-    ssovProperties === undefined ||
-    ssovData === undefined ||
-    selectedSsov === null
-  )
+  if (ssovData === undefined || ssovEpochData === undefined)
     return (
       <Box className="overflow-x-hidden bg-black h-screen">
         <PageLoader />
@@ -70,29 +35,38 @@ const Manage = () => {
     );
 
   return (
-    <Box className="overflow-x-hidden bg-black h-screen">
-      <Head>
-        <title>SSOV | Dopex</title>
-      </Head>
-      <AppBar active="SSOV" />
-      <Box className="py-12 lg:max-w-5xl md:max-w-3xl sm:max-w-xl max-w-md mx-auto px-4 lg:px-0">
-        <Box className="flex flex-col mt-20">
-          <Box className="flex md:flex-row flex-col mb-4 md:justify-between items-center md:items-start">
-            <Description
-              ssovProperties={ssovProperties}
-              ssovData={ssovData}
-              userSsovData={userSsovData}
-            />
-            <ManageCard ssovProperties={ssovProperties} />
+    <BnbConversionProvider>
+      <Box className="overflow-x-hidden bg-black h-screen">
+        <Head>
+          <title>SSOV | Dopex</title>
+        </Head>
+        <AppBar active="SSOV" />
+        <Box className="py-12 lg:max-w-5xl md:max-w-3xl sm:max-w-xl max-w-md mx-auto px-4 lg:px-0">
+          <Box className="flex flex-col mt-20">
+            <Box className="flex md:flex-row flex-col mb-4 md:justify-between items-center md:items-start">
+              <Description
+                ssovData={ssovData}
+                ssovEpochData={ssovEpochData}
+                ssovUserData={ssovUserData}
+                type={selectedSsov.type}
+              />
+              <ManageCard />
+            </Box>
+            {ssovUserData === undefined ? null : <ExerciseList />}
+            {selectedSsov.type === 'PUT' ? null : <Stats className="mt-4" />}
           </Box>
-          {userSsovData === undefined ? null : (
-            <ExerciseList ssovProperties={ssovProperties} />
-          )}
-          <Stats ssovProperties={ssovProperties} className="mt-4" />
         </Box>
       </Box>
-    </Box>
+    </BnbConversionProvider>
   );
 };
 
-export default Manage;
+const ManagePage = () => {
+  return (
+    <SsovProvider>
+      <Manage />
+    </SsovProvider>
+  );
+};
+
+export default ManagePage;
