@@ -3,14 +3,14 @@ import Box from '@material-ui/core/Box';
 import Input from '@material-ui/core/Input';
 import IconButton from '@material-ui/core/IconButton';
 import ArrowBackIcon from '@material-ui/icons/ArrowBack';
-import { BigNumber, utils as ethersUtils } from 'ethers';
+import { utils as ethersUtils } from 'ethers';
 
 import Dialog from 'components/UI/Dialog';
 import Typography from 'components/UI/Typography';
 import CustomButton from 'components/UI/CustomButton';
 
 import { WalletContext } from 'contexts/Wallet';
-import { SsovContext, SsovProperties } from 'contexts/Ssov';
+import { SsovContext } from 'contexts/Ssov';
 
 import getUserReadableAmount from 'utils/contracts/getUserReadableAmount';
 import formatAmount from 'utils/general/formatAmount';
@@ -23,32 +23,29 @@ export interface Props {
   open: boolean;
   handleClose: () => {};
   strikeIndex: number;
-  ssovProperties: SsovProperties;
 }
 
-const Transfer = ({
-  open,
-  handleClose,
-  strikeIndex,
-  ssovProperties,
-}: Props) => {
+const Transfer = ({ open, handleClose, strikeIndex }: Props) => {
   const {
-    updateSsovData,
-    updateUserSsovData,
-    selectedSsov,
-    ssovDataArray,
-    userSsovDataArray,
+    updateSsovEpochData,
+    updateSsovUserData,
+    ssovEpochData,
+    ssovUserData,
+    ssovData,
+    selectedEpoch,
   } = useContext(SsovContext);
   const { accountAddress, signer } = useContext(WalletContext);
 
-  const [transferAmount, setTransferAmount] = useState(0);
+  const sendTx = useSendTx();
+
+  const [transferAmount, setTransferAmount] = useState('0');
   const [recipient, setRecipient] = useState('');
   const [userEpochStrikeTokenBalance, setUserEpochStrikeTokenBalance] =
-    useState<number>(0);
-  const sendTx = useSendTx();
-  const { selectedEpoch } = ssovProperties;
-  const { epochStrikes } = ssovDataArray[selectedSsov];
-  const { epochStrikeTokens } = userSsovDataArray[selectedSsov];
+    useState<string>('0');
+
+  const { epochStrikes } = ssovEpochData;
+  const { epochStrikeTokens } = ssovUserData;
+
   const strikePrice = getUserReadableAmount(epochStrikes[strikeIndex] ?? 0, 8);
   const epochStrikeToken = epochStrikeTokens[strikeIndex];
 
@@ -73,15 +70,13 @@ const Transfer = ({
 
   const updateUserEpochStrikeTokenBalance = useCallback(async () => {
     if (!epochStrikeToken || !accountAddress) {
-      setUserEpochStrikeTokenBalance(0);
+      setUserEpochStrikeTokenBalance('0');
       return;
     }
     const userEpochStrikeTokenBalance = await epochStrikeToken.balanceOf(
       accountAddress
     );
-    setUserEpochStrikeTokenBalance(
-      getUserReadableAmount(userEpochStrikeTokenBalance, 18)
-    );
+    setUserEpochStrikeTokenBalance(userEpochStrikeTokenBalance.toString());
   }, [epochStrikeToken, accountAddress]);
 
   const handleRecipientChange = useCallback((e) => {
@@ -100,12 +95,14 @@ const Transfer = ({
     if (!accountAddress || !epochStrikeToken) return;
     try {
       sendTx(
-        epochStrikeToken.connect(signer).transfer(accountAddress, recipient)
+        epochStrikeToken
+          .connect(signer)
+          .transfer(recipient, ethersUtils.parseEther(String(transferAmount)))
       );
-      updateSsovData();
-      updateUserSsovData();
+      updateSsovEpochData();
+      updateSsovUserData();
       updateUserEpochStrikeTokenBalance();
-      setTransferAmount(0);
+      setTransferAmount('0');
       setRecipient('');
     } catch (err) {
       console.log(err);
@@ -114,16 +111,18 @@ const Transfer = ({
     accountAddress,
     epochStrikeToken,
     recipient,
-    updateSsovData,
+    updateSsovEpochData,
     updateUserEpochStrikeTokenBalance,
-    updateUserSsovData,
+    updateSsovUserData,
     signer,
     sendTx,
+    transferAmount,
   ]);
 
   useEffect(() => {
     updateUserEpochStrikeTokenBalance();
   }, [updateUserEpochStrikeTokenBalance]);
+
   return (
     <Dialog
       open={open}
@@ -142,7 +141,7 @@ const Transfer = ({
         </Box>
         <Box className="flex justify-between">
           <Typography variant="h6" className="text-stieglitz mb-2">
-            Balance: {userEpochStrikeTokenBalance.toString()}
+            Balance: {getUserReadableAmount(userEpochStrikeTokenBalance, 18)}
           </Typography>
           <Typography
             variant="h6"
@@ -158,12 +157,12 @@ const Transfer = ({
             <Box className="h-12 bg-cod-gray rounded-xl p-2 flex flex-row items-center">
               <Box className="flex flex-row h-8 w-8 mr-2">
                 <img
-                  src={SSOV_MAP[ssovProperties.tokenName].imageSrc}
-                  alt={ssovProperties.tokenName}
+                  src={SSOV_MAP[ssovData.tokenName].imageSrc}
+                  alt={ssovData.tokenName}
                 />
               </Box>
               <Typography variant="h5" className="text-white">
-                {ssovProperties.tokenName}
+                {ssovData.tokenName}
               </Typography>
             </Box>
             <Input
@@ -225,11 +224,7 @@ const Transfer = ({
           className="w-full mb-4"
           onClick={handleTransfer}
           size="xl"
-          disabled={
-            transferAmount > 0 && recipient !== '' && error === undefined
-              ? false
-              : true
-          }
+          disabled={recipient !== '' && error === undefined ? false : true}
         >
           Transfer
         </CustomButton>
