@@ -7,47 +7,49 @@ import Countdown from 'react-countdown';
 import CustomButton from 'components/UI/CustomButton';
 import Typography from 'components/UI/Typography';
 
-import { SsovContext, SsovProperties } from 'contexts/Ssov';
+import { SsovContext } from 'contexts/Ssov';
 import { AssetsContext } from 'contexts/Assets';
 
 import getUserReadableAmount from 'utils/contracts/getUserReadableAmount';
 import formatAmount from 'utils/general/formatAmount';
+
 import { SSOV_MAP } from 'constants/index';
 
 import useSendTx from 'hooks/useSendTx';
 
 import styles from './styles.module.scss';
 
-const Withdraw = ({ ssovProperties }: { ssovProperties: SsovProperties }) => {
+const Withdraw = () => {
   const {
-    updateSsovData,
-    updateUserSsovData,
+    updateSsovEpochData,
+    updateSsovUserData,
+    ssovEpochData,
+    ssovData,
+    ssovUserData,
+    ssovSigner,
+    selectedEpoch,
     selectedSsov,
-    ssovDataArray,
-    userSsovDataArray,
-    ssovSignerArray,
   } = useContext(SsovContext);
 
-  const { selectedEpoch, tokenName } = ssovProperties;
-  const { ssovContractWithSigner } = ssovSignerArray[selectedSsov];
+  const { tokenName } = ssovData;
+  const { ssovContractWithSigner } = ssovSigner;
   const {
     epochTimes,
     epochStrikes,
     totalEpochStrikeDeposits,
     totalEpochDeposits,
-  } = ssovDataArray[selectedSsov];
-  const { userEpochStrikeDeposits, userEpochDeposits } =
-    userSsovDataArray[selectedSsov];
+  } = ssovEpochData;
+  const { userEpochStrikeDeposits, userEpochDeposits } = ssovUserData;
 
   const { updateAssetBalances } = useContext(AssetsContext);
 
   const sendTx = useSendTx();
 
-  // Ssov data for next epoch
-
   const epochEndTime = epochTimes[1]
     ? format(new Date(epochTimes[1] * 1000), 'MM/dd')
     : 'N/A';
+
+  const isPut = selectedSsov.type === 'PUT';
 
   const strikes = epochStrikes.map((strike) =>
     getUserReadableAmount(strike, 8).toString()
@@ -77,15 +79,15 @@ const Withdraw = ({ ssovProperties }: { ssovProperties: SsovProperties }) => {
       ? getUserReadableAmount(userEpochDeposits, 8)
       : getUserReadableAmount(userEpochDeposits, 18);
 
-  const tokenSymbol = SSOV_MAP[ssovProperties.tokenName].tokenSymbol;
+  const tokenSymbol = SSOV_MAP[ssovData.tokenName].tokenSymbol;
 
   // Handle Withdraw
   const handleWithdraw = useCallback(
     async (index) => {
       try {
         await sendTx(ssovContractWithSigner.withdraw(selectedEpoch, index));
-        updateSsovData();
-        updateUserSsovData();
+        updateSsovEpochData();
+        updateSsovUserData();
       } catch (err) {
         console.log(err);
       }
@@ -94,8 +96,8 @@ const Withdraw = ({ ssovProperties }: { ssovProperties: SsovProperties }) => {
     [
       ssovContractWithSigner,
       selectedEpoch,
-      updateSsovData,
-      updateUserSsovData,
+      updateSsovEpochData,
+      updateSsovUserData,
       updateAssetBalances,
       sendTx,
     ]
@@ -112,7 +114,8 @@ const Withdraw = ({ ssovProperties }: { ssovProperties: SsovProperties }) => {
             <span className="text-wave-blue">
               {formatAmount(userEpochDepositsAmount, 5)}
             </span>{' '}
-            / {formatAmount(totalEpochDepositsAmount, 5)} {tokenSymbol}
+            / {formatAmount(totalEpochDepositsAmount, 5)}{' '}
+            {isPut ? '2CRV' : tokenSymbol}
           </Typography>
         </Box>
         <Box>
@@ -154,7 +157,8 @@ const Withdraw = ({ ssovProperties }: { ssovProperties: SsovProperties }) => {
                 key={index}
               >
                 <Typography variant="h6">
-                  {formatAmount(totalEpochStrikeDepositsAmounts[index], 5)}
+                  {formatAmount(totalEpochStrikeDepositsAmounts[index], 5)}{' '}
+                  {isPut ? '2CRV' : tokenSymbol}
                 </Typography>
                 <Typography variant="h6" className="text-stieglitz">
                   {tokenSymbol} ${strike}
